Fix rebase

+++ lib/Dialect/TritonGPU/IR/LinearLayoutConversions.cpp
@@ -176,7 +176,7 @@ LinearLayout ensureLayoutNotLargerThan(
           continue;
         }
         assert(llvm::isPowerOf2_32(outValue));
-        sortedBases.emplace_back(inDimName, basisIdx, outValue);
+        sortedBases.emplace_back(inDimName, (int)basisIdx, outValue);
       }
     }
     // From the largest basis to the smallest.
